refactor(pangraph): move sequence reconstruction into the graph module
