+++ packages/pangraph/src/pangraph/mod.rs
@@ -6,6 +6,7 @@ pub mod pangraph_block;
 pub mod pangraph_interval;
 pub mod pangraph_node;
 pub mod pangraph_path;
+pub mod reconstruct;
 pub mod reweave;
 pub mod slice;
 pub mod split_matches;
